Brought our delete-order router a bit more up to speed with our NATS flow

+++ orders/src/routes/delete-order.ts
@@ -1,6 +1,12 @@
 import express, { Request, Response } from "express";
-import { requireAuth, validateRequest } from "@cream-paws-util/common";
+import {
+   NotFoundError,
+   requireAuth,
+   validateRequest,
+} from "@cream-paws-util/common";
 import { Order } from "../models/order";
+import { OrderDeletedPublisher } from "../events/publishers/order-deleted-publisher";
+import { natsWrapper } from "../nats-wrapper";
 
 const router = express.Router();
 
@@ -8,8 +14,22 @@ router.delete(
    "/api/orders/:id",
    [requireAuth, validateRequest],
    async (req: Request, res: Response) => {
+      const { id } = req.params;
+
+      const order = await Order.findById(id);
+
+      if (!order) {
+         throw new NotFoundError();
+      }
+
       await Order.deleteOne({
-         id: req.params.id,
+         id,
+      });
+
+      new OrderDeletedPublisher(natsWrapper.client).publish({
+         id,
+         customerId: order.customer_id,
+         version: order.version,
       });
 
       res.status(200).send(`Order id: ${req.params.id} successfully deleted.`);
